+++ src/modules/dashboard/Board/components/Dialogs/DialogEditBoard/components/DialogBackground/services/types/fetchGetSearchImageService.types.ts
